numeric not integer for number columns

+++ Model/src/main/java/org/gusdb/wdk/model/query/ColumnType.java
@@ -68,7 +68,7 @@ public enum ColumnType {
       case CLOB:
         return "text";
       case NUMBER:
-        return "integer";
+        return "numeric";
       case FLOAT:
         return "double precision";
       case BOOLEAN:
